python/requirements.txt
numpy==2.4.1
pillow==12.1.0
zxing-cpp==2.3.0
